+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Unreleased (Use snapshot builds to get these changes)
 * HM-CC-RT-DN: Added specific attributes/methods: ``device.temperature``,
   ``device.set_temperature``, ``device.is_off``, ``device.turn_off()``,
   ``device.control_mode``, ``device.battery_state``
+* HM-PBI-4-FM: Changed access to switches from ``device.button(0)`` to
+  a hopefully clearer ``device.switch1``, ``device.switch2``, ...
 
 Version 0.1 (2016-03-13)
 ------------------------